use int

+++ db/migrate/20260227000000_add_compute_time_to_images_to_video_tasks.rb
@@ -1,5 +1,5 @@
 class AddComputeTimeToImagesToVideoTasks < ActiveRecord::Migration[8.1]
   def change
-    add_column :images_to_video_tasks, :completion_time, :float
+    add_column :images_to_video_tasks, :completion_time, :integer
   end
 end
